Regenerate documentation images using pekko implementations.
Using this command (src https://stackoverflow.com/a/24879385)
I found all the images in the repo ...
```
find . -type f -exec file --mime-type {} \+ | awk -F: '{if ($2 ~/image\//) print $1}'
```

1. DispatcherBehaviourOnBadCode.png & DispatcherBehaviourOnGoodCode.png

These are used in `handling-blocking-operations-in-pekko-http-routes.md`
to document the effect of blocking operations.

To regenerate these ...
 * I created a sample project using pekko libraries.
 * I used the routes documented in the article.
 * I concatenated a `get` route to allow for traffic with GET requests.
 * I used the following script trying to follow along with the traffic
   the article was describing ...
```
for i in $(seq 256);
do
  sleep 0.05
  curl http://localhost:8080 &
done
wait
for i in $(seq 256);
do
  sleep 0.05
  curl -X POST http://localhost:8080 &
  sleep 0.05
  curl http://localhost:8080 &
done
```
 * I started a recording from Java Mission Control.
 * I viewed that recording in the Legacy Threads view.
 * I clicked on the default or blocking dispatcher threads.
 * I stored and set that selection.
 * I then went into the new Threads view and zoomed in.

I changed the language in the doc to match my recording setup.
That language includes mention of color and the tools used.

In trying out the code of the article I found I needed changes ...
I've estimated what the proper fix is for the Java example,
though I haven't checked it in my example app.

```
[error] /Users/[...]]/Desktop/pekko-example/src/main/scala/com/example/QuickstartApp.scala:20:65: type mismatch;
[error]  found   : String("my-blocking-dispatcher")
[error]  required: org.apache.pekko.actor.typed.DispatcherSelector
```

This code was committed over 7 years ago
4ae775007f29ad2a13da37e7150e7d665087583e
and I anticipate the API has changed since then.

2. DispatcherBehaviourProperBlocking.png

This image isn't in use today ...

```
% git grep DispatcherBehaviourProperBlocking.png
```

And has never been used ...

```
% git log -S "DispatcherBehaviourProperBlocking"
```

Not even when the file was first checked in
4ae775007f29ad2a13da37e7150e7d665087583e.

3. akka-http-file-listing.png

This one was easy enough.
I added a route to return `getFromBrowsableDirectory`.
I pointed that to some example directory I created,
and took a screenshot from Safari.