Improve too large chunk size test (#529)
The current test goes over the configured max chunk size when parsing
the last hex char. The spirit of the code is to continue parsing the
size even if we've already gone over the max chunk size. To actually
test this we need to make sure another char follows after we've gone
over the max chunk size.